+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Show uniq request id in API HTTP response. (#425, @miry)
 * Add method to parse `stream.Direction` from string.
   Allow to convert `stream.Direction` to string. (#430, @miry)
+* Add posibility to write to Output with deadline.
+  On interrupting badnwidth toxic use non blocking write. (#436, @miry)
 
 # [2.4.0] - 2022-03-07